feat: 인증 인가, 승인 메세지 전송, 보상 반영 기능 구현

+++ src/main/java/com/road_journey/road_journey/goals/dto/GoalRewardResponseDto.java
@@ -6,8 +6,8 @@ import lombok.Getter;
 @Getter
 public class GoalRewardResponseDto {
     private final boolean isReward;
-    private final int gold;
-    private final int growthPoint;
+    private final Long gold;
+    private final Long growthPoint;
 
     public GoalRewardResponseDto(Goal goal) {
         this.isReward = goal.isLastPeriodCompleted();
